+++ project2.sh
@@ -44,3 +44,6 @@ fi
 #$VENV/bin/vallm batch --parallel .
 $VENV/bin/vallm batch . --recursive --format toon --output ./project --exclude "CHANGELOG.md" --exclude "*.md"
 $VENV/bin/prefact -a -e "examples/**"
+$PIP install sumd --upgrade --quiet
+$VENV/bin/sumd .
+$VENV/bin/sumr .